requirements.txt
requests==2.31.0
jinja2==3.1.2
markdown==3.5.1
cryptography==42.0.5
numpy>=2.1.0,<3
faiss-cpu==1.13.2
sentence-transformers>=3.0.0
